CR - Refactor useCarouselState
