+++ apps/filecoin-site/src/app/_components/Carousel/Carousel.tsx
@@ -1,12 +1,6 @@
 'use client'
 
-import {
-  useState,
-  useCallback,
-  useEffect,
-  createContext,
-  useContext,
-} from 'react'
+import { createContext, useContext } from 'react'
 
 import { clsx } from 'clsx'
 import useEmblaCarousel, {
@@ -14,6 +8,7 @@ import useEmblaCarousel, {
 } from 'embla-carousel-react'
 
 import { useCarouselKeyboard } from './hooks/useCarouselKeyboard'
+import { useCarouselState } from './hooks/useCarouselState'
 
 export type CarouselApi = UseEmblaCarouselType[1]
 type UseCarouselParameters = Parameters<typeof useEmblaCarousel>
@@ -64,40 +59,10 @@ export function Carousel({
     },
     plugins,
   )
-  const [canScrollPrev, setCanScrollPrev] = useState(false)
-  const [canScrollNext, setCanScrollNext] = useState(false)
 
-  const onSelect = useCallback((api: CarouselApi) => {
-    if (!api) return
-    setCanScrollPrev(api.canScrollPrev())
-    setCanScrollNext(api.canScrollNext())
-  }, [])
-
-  const scrollPrev = useCallback(() => {
-    api?.scrollPrev()
-  }, [api])
-
-  const scrollNext = useCallback(() => {
-    api?.scrollNext()
-  }, [api])
-
+  const { canScrollPrev, canScrollNext, scrollPrev, scrollNext } =
+    useCarouselState(api, setApi)
   const handleKeyDown = useCarouselKeyboard(scrollPrev, scrollNext)
-
-  useEffect(() => {
-    if (!api || !setApi) return
-    setApi(api)
-  }, [api, setApi])
-
-  useEffect(() => {
-    if (!api) return
-    onSelect(api)
-    api.on('reInit', onSelect)
-    api.on('select', onSelect)
-
-    return () => {
-      api?.off('select', onSelect)
-    }
-  }, [api, onSelect])
 
   return (
     <CarouselContext.Provider
@@ -126,6 +91,6 @@ export function Carousel({
   )
 }
 
-function getCarouselAxis(orientation: 'horizontal' | 'vertical') {
+function getCarouselAxis(orientation: CarouselProps['orientation']) {
   return orientation === 'horizontal' ? 'x' : 'y'
 }
