+++ apps/filecoin-site/src/app/_components/Carousel/hooks/useCarouselState.ts
@@ -1,0 +1,48 @@
+import { useState, useCallback, useEffect } from 'react'
+
+import { type CarouselApi } from '../Carousel'
+
+export function useCarouselState(
+  api: CarouselApi,
+  setApi?: (api: CarouselApi) => void,
+) {
+  const [canScrollPrev, setCanScrollPrev] = useState(false)
+  const [canScrollNext, setCanScrollNext] = useState(false)
+
+  const onSelect = useCallback((api: CarouselApi) => {
+    if (!api) return
+    setCanScrollPrev(api.canScrollPrev())
+    setCanScrollNext(api.canScrollNext())
+  }, [])
+
+  const scrollPrev = useCallback(() => {
+    api?.scrollPrev()
+  }, [api])
+
+  const scrollNext = useCallback(() => {
+    api?.scrollNext()
+  }, [api])
+
+  useEffect(() => {
+    if (!api || !setApi) return
+    setApi(api)
+  }, [api, setApi])
+
+  useEffect(() => {
+    if (!api) return
+    onSelect(api)
+    api.on('reInit', onSelect)
+    api.on('select', onSelect)
+
+    return () => {
+      api?.off('select', onSelect)
+    }
+  }, [api, onSelect])
+
+  return {
+    canScrollPrev,
+    canScrollNext,
+    scrollPrev,
+    scrollNext,
+  }
+}
